Project Navigation updated design changes

+++ app/Http/Controllers/ProjectController.php
@@ -30,9 +30,17 @@ class ProjectController extends Controller
      */
     public function index()
     {
+        $projects = DB::table('projects')
+       /* ->join('users', 'users.id', '=', 'projects.assigned_to')
+        ->join('customers', 'customers.id', '=', 'projects.customer_id')*/
+        ->get();
+        //dd($projects);
+        //$projectstotal = DB::table('projects')->get();//Project::all();
+        $contractors = DB::table('users')->where('role_id', '=', '1')->get();
+        return view('projects/index', compact('projects'),['contractors' => $contractors]);
+        //return view('projects.index');
 
-        return view('projects.index');
-    }
+    } 
 
     /**
      * Show the form for creating a new resource.
@@ -41,8 +49,8 @@ class ProjectController extends Controller
      */
     public function create()
     {
-        $contractors = DB::table('users')->where('role_id', '=', '3')->get();
-        return View::make('projects/create')->with('contractors', $contractors);
+        $contractors = DB::table('users')->where('role_id', '=', '1')->get();
+        return View('projects/create')->with('contractors', $contractors);
         //return view('projects/create');
     }
 
@@ -81,7 +89,7 @@ class ProjectController extends Controller
         // dd($contractor[0]->id);
 
         $customer = new Customer([
-            'name' => $request->input('name'),
+            'c_name' => $request->input('name'),
             'cnic' => $request->input('cnic'),
             'phone' => $request->input('phone'),
             'address' => $request->input('address'),
@@ -103,7 +111,7 @@ class ProjectController extends Controller
             'plot_size' => $request->input('plot_size'),
             'floor' => $request->input('floor'),
             'customer_id' => $customerId[0]->id,
-            'assigned_to' => $contractor[0]->id,
+            'assigned_to' =>DB::table('users')->where('name','=', $request->input('assigned_to'))->pluck('id')->first(),
             'status_id' => $project_status,
             'phase_id' => $project_phase,
             'assigned_by' => Auth::id(),
@@ -111,6 +119,7 @@ class ProjectController extends Controller
             'estimated_budget' => $request->input('estimated_budget'),
             'description' => $request->input('description'),
             'contract_image' => $request->input('contract_image')
+            
 
         ]);
 
@@ -145,14 +154,17 @@ class ProjectController extends Controller
      * @param \App\Project $project
      * @return \Illuminate\Http\Response
      */
+
     public function show()
     {
         $projects = DB::table('projects')
-        ->join('users', 'users.id', '=', 'projects.assigned_to')
-        ->join('customers', 'customers.id', '=', 'projects.customer_id')
+        /*->join('users', 'users.id', '=', 'projects.assigned_to')
+        ->join('customers', 'customers.id', '=', 'projects.customer_id')*/
         ->get();
-        dd($projects);
-//->where('followers.follower_id', '=', 3)
+
+        
+        //dd($projects);
+        //->where('followers.follower_id', '=', 3)
        // $projects = DB::table('projects')->join('users','users.id','=','projects.assigned_to')->take(10)->get();
         //dd($projects);
           //Project::paginate(10);
@@ -164,8 +176,38 @@ class ProjectController extends Controller
         //dd($customers);
         //join('customers','customers.id','=','projects.customer_id')->take(10)->get();
         //$customers = DB::table('customers')->get();// \App\Customer::all();
-        return view('projects/index', compact('projects'));
+        return view('projects/projects', compact('projects'));
     }
+
+
+    public function cancelled()
+    {
+      $projects = DB::table('projects')->where('projects.id','=','5')->get();
+      return view('projects/projects',compact('projects'));
+    }
+
+    public function completed()
+    {
+     $projects = DB::table('projects')->where('projects.floor','=','5')->get();
+      return view('projects/projects',compact('projects'));
+    }
+    public function all()
+    {
+     $projects = DB::table('projects')->get();
+      return view('projects/projects',compact('projects'));
+    }
+     public function pending()
+    {
+     $projects = DB::table('projects')->where('projects.floor','=','3')->get();
+      return view('projects/projects',compact('projects'));
+    }
+     public function current()
+    {
+     $projects = DB::table('projects')->where('projects.floor','=','3')->get();
+      return view('projects/projects',compact('projects'));
+    }
+
+
 
     /**
      * Show the form for editing the specified resource.
@@ -193,27 +235,49 @@ class ProjectController extends Controller
 
     public function edit($id)
     {
+
         $projects = Project::find($id);
         // Redirect to user list if updating user wasn't existed
         if ($projects == null || count($projects) == 0) {
             return redirect()->intended('projects/index');
         }
         //$users = User::paginate(10);
-        return view('projects/view', ['projects' => $projects]);
+        return view('projects/edit', ['projects' => $projects]);
     }
-    /*
+
 
      public function viewuser($id)
     {
+        $labors = DB::table('labors')
+        ->join('projects', 'projects.id', '=', 'labors.project_id')
+        ->get();
+
+       /* $customers = DB::table('customers')
+        ->join('projects', 'projects.customer_id', '=', 'customers.id')
+        ->get();
+        */
+        $orders = DB::table('order_details')->join('projects','projects.id','=','order_details.project_id');
+        $customers = DB::table('customers')->where('projects.id', '=', $id)
+        ->join('projects', 'projects.customer_id', '=', 'customers.id')
+        ->get();
+       
+        $data = DB::table('projects')->where('projects.id', '=', $id)
+        ->join('users', 'users.id', '=', 'projects.assigned_to')
+        ->join('customers', 'customers.id', '=', 'projects.customer_id')
+        ->get();
+        
+
         $projects = Project::find($id);
         // Redirect to user list if updating user wasn't existed
         if ($projects == null || count($projects) == 0) {
             return redirect()->intended('projects/index');
         }
+
         //$users = User::paginate(10);
-        return view('projects/viewuser', ['projects' => $projects]);
+        return view('projects/view', ['data' => $data] , ['labors' => $labors], ['orders' => $orders ]);
     }
-*/
+
+
 
     /**
      * Update the specified resource in storage.
@@ -263,9 +327,24 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
-        Project::where('id', $id)->delete();
-        return redirect()->intended('projects/index');
+       
+        $check = DB::table('labors')
+        ->join('projects', 'projects.id', '=', 'labors.project_id') //->where('labors.id','=',$id)
+        ->count();
+      
+        if($check == 0)
+        {
+            Project::where('id', $id)->delete();
+            return redirect()->intended('projects/index')->with('success','Project Deleted Succuessfully.');
+        }
+        else
+        {
+            return redirect()->route('projects.index')->with('message', 'Labor Exists. Please delete labors from project first.');
+
+        }        
     }
+
+
 
 
     public function updateImage(Request $request)
